refactor: UiUtils를 Utils로 변경 및 GameUtils에서 일부 메소드 분리
ex) generateRandomNumbors과 같은 lottoNumber를 직접적으로 사용하지는 않으나 필요한 부분들
