+++ __tests__/GameUtilsTest.js
@@ -7,19 +7,6 @@ describe('GameUtils 클래스', () => {
     new Lotto([7, 8, 9, 10, 11, 12]),
   ];
 
-  describe('generateRandomNumbers 메소드', () => {
-    test('1~45까지 중복되지 않는 길이가 6인 랜덤 숫자 배열을 반환.', () => {
-      const gameUtils = new GameUtils(mockLottos);
-      const randomNumbers = gameUtils.generateRandomNumbers();
-      expect(randomNumbers.length).toBe(mockLottos[0].getNumbers().length);
-      expect(new Set(randomNumbers).size).toBe(randomNumbers.length);
-      randomNumbers.forEach((num) => {
-        expect(num).toBeGreaterThanOrEqual(1);
-        expect(num).toBeLessThanOrEqual(45);
-      });
-    });
-  });
-
   describe('matchingCountsWithArr 메소드', () => {
     test('당첨 번호를 기반으로 각 로또의 일치하는 숫자 수를 배열로 반환', () => {
       const gameUtils = new GameUtils(mockLottos);
@@ -42,33 +29,6 @@ describe('GameUtils 클래스', () => {
     });
   });
 
-  describe('addOrUpdatePropertyInObj 메소드', () => {
-    test('객체에 속성을 추가 또는 업데이트 하고 반환.', () => {
-      const gameUtils = new GameUtils(mockLottos);
-      const obj = { 2: 1, 4: 2 };
-      const newObj = gameUtils.addOrUpdatePropertyInObj(obj, 4);
-      expect(newObj).toEqual({ 2: 1, 4: 3 });
-    });
-  });
-
-  describe('removeItemsWithNumericKeysLessThanThree 메소드', () => {
-    test('숫자 키가 3보다 작은 항목을 객체에서 제거.', () => {
-      const gameUtils = new GameUtils(mockLottos);
-      const obj = { 1: 2, 2: 3, 3: 4, 4: 5 };
-      const newObj = gameUtils.removeItemsWithNumericKeysLessThanThree(obj);
-      expect(newObj).toEqual({ 3: 4, 4: 5 });
-    });
-  });
-
-  describe('addMissingElements 메소드', () => {
-    it('WINNIG_PROFITS의 키와 일치하는 속성이 객체에 없으면 0으로 초기화해야 합니다.', () => {
-      const gameUtils = new GameUtils(mockLottos);
-      const obj = { 2: 1, 4: 2 };
-      const newObj = gameUtils.addMissingElements(obj);
-      expect(newObj).toEqual({ 2: 1, 4: 2, 3: 0, 5: 0,'bonus':0, 6: 0 }); // 예시 값에 따라 수정
-    });
-  });
-  
   describe('processMatchingNumbersToResult 메소드', () => {
     test('일치하는 숫자의 개수를 모아놓은 객체를 당첨결과로 볼 수 있도록 수정 (addMissingElements, removeItemsWithNumericKeysLessThanThree)', () => {
       const gameUtils = new GameUtils(mockLottos);
@@ -90,6 +50,4 @@ describe('GameUtils 클래스', () => {
       expect(result2).toBe('5');
     });
   });
-
-  // 나머지 메소드에 대한 테스트도 유사한 방식으로 작성 가능합니다.
 });
