+++ src/ui/OutputUi.js
@@ -1,6 +1,5 @@
 import { MissionUtils } from '@woowacourse/mission-utils';
 import { OUTPUT_MESSAGE, WINNING_NUMBERS } from '../Constants.js';
-import UiUtils from './UiUtils.js';
 
 class OutputUi {
   constructor() {}
@@ -23,7 +22,7 @@ class OutputUi {
     for (const index in WINNING_NUMBERS) {
       const MATHING_COUNT = WINNING_NUMBERS[index];
       MissionUtils.Console.print(
-        UiUtils.winningStatusForUser(MATHING_COUNT, winnigStatus)
+        OUTPUT_MESSAGE.winningStatus(MATHING_COUNT, winnigStatus)
       );
     }
   }
